fix order specs

+++ db/migrate/20140102195834_add_total_price_to_cart.rb
@@ -1,5 +1,5 @@
 class AddTotalPriceToCart < ActiveRecord::Migration
   def change
-    add_column :carts, :total_price, :int
+    add_column :carts, :total_price, :float
   end
 end
